requirements.txt
pydantic==1.8.2
fastapi==0.68.1
mysql_connector_repackaged==0.3.1
python-dotenv==0.19.0
uvicorn==0.13.4
